pyproject.toml
[tool.poetry]
name = "automsreward"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.48.0"
httpx = "^0.27.0"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
schedule = "^1.2.1"
pre-commit = "^3.7.0"
python-telegram-bot = "^21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
